+++ src/main/java/com/phantomwing/rusticdelight/datagen/ModBlockStateProvider.java
@@ -13,6 +13,7 @@ import net.minecraftforge.client.model.generators.ConfiguredModel;
 import net.minecraftforge.client.model.generators.ModelFile;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.registries.ForgeRegistries;
+import vectorwing.farmersdelight.FarmersDelight;
 import vectorwing.farmersdelight.common.block.PieBlock;
 
 import java.util.function.Function;
@@ -98,7 +99,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
     private ConfiguredModel[] bellPepperStates(BlockState state, CropBlock block, String modelName, String textureName) {
         var ageProperty = state.getValue(((BellPepperCropBlock) block).getAgeProperty());
         ConfiguredModel[] models = new ConfiguredModel[1];
-        models[0] = new ConfiguredModel(models().cross(modelName + ageProperty,
+        models[0] = new ConfiguredModel(models().singleTexture(modelName + ageProperty, resourceBlock("crop_cross"), "cross",
                 new ResourceLocation(RusticDelight.MOD_ID, "block/" + textureName + ageProperty)).renderType("cutout"));
 
         return models;
@@ -113,7 +114,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
     private ConfiguredModel[] coffeeStates(BlockState state, CropBlock block, String modelName, String textureName) {
         var ageProperty = state.getValue(((CoffeeCropBlock) block).getAgeProperty());
         ConfiguredModel[] models = new ConfiguredModel[1];
-        models[0] = new ConfiguredModel(models().cross(modelName + ageProperty,
+        models[0] = new ConfiguredModel(models().singleTexture(modelName + ageProperty, resourceBlock("crop_cross"), "cross",
                 new ResourceLocation(RusticDelight.MOD_ID, "block/" + textureName + ageProperty)).renderType("cutout"));
 
         return models;